casts needed for i686 glibc
git-svn-id: https://svn.r-project.org/R/trunk@86272 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/util.c
@@ -1555,7 +1555,7 @@ size_t wcstoutf8(char *s, const wchar_t *wc, size_t n)
 	    p++;
 	} else {
 	    if (IS_HIGH_SURROGATE(*p) || IS_LOW_SURROGATE(*p))
-		warning("unpaired surrogate Unicode point %x", *p);
+		warning("unpaired surrogate Unicode point %x", (unsigned int)*p);
 	    m = Rwcrtomb32(t, (R_wchar_t)(*p), n - res);
 	}
 	if (!m) break;
